Refactor: ChordsEditor moved to 'common' folder + alias for 'common' folder
