+++ src/components/common/ChordsEditor/Chord/Chord.scss
