+++ src/components/common/ChordsEditor/Chord/index.ts
