+++ src/components/common/ChordsEditor/ChordEditPopup/ChordEditPopup.scss
