+++ src/components/common/ChordsEditor/ChordEditPopup/index.ts
