+++ src/components/common/ChordsEditor/EditableHeader/EditableHeader.scss
