+++ src/components/common/ChordsEditor/EditableHeader/index.ts
