+++ src/components/common/ChordsEditor/InlineInput/InlineInput.scss
