+++ src/components/common/ChordsEditor/InlineInput/index.ts
