+++ src/components/common/ChordsEditor/Letter/Letter.scss
