+++ src/components/common/ChordsEditor/Letter/index.ts
