+++ src/components/common/ChordsEditor/index.ts
